Basic reduction of JS files
- grouped related resources toghether to decrease coupling while
resources in files still maintain an overall high cohesion
