--- assets/colours/defaultColours.js
@@ -1,3 +1,0 @@
-export const DefaultColours = {
-    None: "none"
-};
