--- assets/strings/ageSexGroupHeadings.js
@@ -1,7 +1,0 @@
-export const AgeSexGroupHeadings = {
-    Population1Up: "Population age 1+", 
-    Children1To8: "Children 1 to 8 y", 
-    YouthAndAdolescents: "Youth & adolescents 9 to 18 y",
-    AdultMales: "Adult males 19 y +",
-    AdultFemales: "Adult females 19 y +"
-};
