--- assets/strings/attributes.js
@@ -1,17 +1,0 @@
-export const TextWrap  = {
-    NoWrap: "No Wrap",
-    Wrap: "Wrap"
-};
-
-export const TextAnchor = {
-    Start: "start",
-    Middle: "middle",
-    End: "end"
-};
-
-export const FontWeight = {
-    Normal: "normal",
-    Bold: "bold",
-    Lighter: "lighter",
-    Bolder: "bolder"
-}
