--- assets/strings/columnNames.js
@@ -1,13 +1,0 @@
-export const FoodGroupDescDataColNames = {
-    foodGroupLv1: "Food Group Name_Level 1",
-    foodGroupLv2: "Food Group Name_Level 2",
-    foodGroupLv3: "Food Group Name_Level 3",
-    description: "Description of the Contents of the Food Groups and Sub-groups"
-};
-
-export const NutrientDataColNames = {
-    ageSexGroup: "Age-sex group (*: excludes pregnant or breastfeeding)",
-    foodGroupLv1: "Food group_level1",
-    foodGroupLv2: "Food group_level2",
-    foodGroupLv3: "Food group_level3"
-};
