--- assets/strings/defaultAttributes.js
@@ -1,8 +1,0 @@
-import { TextAnchor, TextWrap, FontWeight} from "./attributes.js"
-
-export const DefaultAttributes = {
-    textAnchor: TextAnchor.Start,
-    textWrap: TextWrap.Wrap,
-    fontWeight: FontWeight.Normal,
-    opacity: 1
-}
