+++ index.html
@@ -654,7 +654,7 @@
 
     <script src="i18next.js"></script>
     <script src="i18nextBrowserLanguageDetector.min.js"></script>
-    <script type="module" src="./viewController/foodSourceContributionTool.js"></script>
+    <script type="module" src="./viewController/viewController.js"></script>
     <script>
         jQuery.fn.tableToCSV = function () {
             let clean_text = function (text) {
