--- model/enums/ageSexGroupOrder.js
@@ -1,9 +1,0 @@
-import { AgeSexGroupHeadings } from "../../assets/strings/ageSexGroupHeadings.js";
-
-export const AgeSexGroupOrder = {};
-
-AgeSexGroupOrder[AgeSexGroupHeadings.Population1Up] = 0;
-AgeSexGroupOrder[AgeSexGroupHeadings.Children1To8] = 1;
-AgeSexGroupOrder[AgeSexGroupHeadings.YouthAndAdolescents] = 2;
-AgeSexGroupOrder[AgeSexGroupHeadings.AdultMales] = 3;
-AgeSexGroupOrder[AgeSexGroupHeadings.AdultFemales] = 4;
