--- viewController/enums/sunBurstStates.js
@@ -1,4 +1,0 @@
-export const SunBurstStates = {
-    AllDisplayed: "All Displayed",
-    FilterOnlyLevel2: "Filter Only Level 2"
-};
